requirements.txt
aniso8601==1.3.0
cachetools==2.0.1
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
dill==0.2.7.1
Flask==0.12.2
Flask-Jsonpify==1.5.0
Flask-RESTful==0.3.6
future==0.16.0
google-auth==1.1.1
google-cloud-core==0.27.1
google-cloud-language==0.29.0
google-gax==0.15.15
googleapis-common-protos==1.5.3
grpcio==1.4.0
gunicorn==19.7.1
idna==2.6
itsdangerous==0.24
Jinja2==2.9.6
MarkupSafe==1.0
numpy==1.13.3+mkl
oauthlib==2.0.4
pandas==0.20.3
ply==3.8
protobuf==3.4.0
pyasn1==0.3.7
pyasn1-modules==0.1.4
pyreadline==2.1
python-dateutil==2.6.1
pytz==2017.2
requests==2.18.4
requests-oauthlib==0.8.0
rsa==3.4.2
scikit-learn==0.19.0
scipy==1.0.0rc1
six==1.11.0
sklearn==0.0
SQLAlchemy==1.1.14
tweepy==3.5.0
urllib3==1.22
Werkzeug==0.12.2
